kernel-fit-image: enable compatible string for kernel built-in DTBs
At present, the kernel-fit-image bbclass is adding compatible strings only
for external DTBs, omitting them for DTBs built from the kernel tree. This
behavior is limiting the ability to clearly identify the DTBs in FIT image
that includes multiple kernel built-in DTBs and possibly mix of in tree and
external ones, due to lack of consistent compatible strings.

To fix this, introduce a new BitBake variable, FIT_DTB_ADD_COMPATIBLE
when set to "True", the bbclass will extract and include the compatible
string from kernel-built DTBs also in the .its file, just like it does
for external DTBs.

Signed-off-by: Kavinaya S <kavinaya@qti.qualcomm.com>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>