oe-selftest: fitimage: test FIT image with/without kernel node
Add test coverage for the feature added with commit
oe-selftest: fitimage: add FIT_LINUX_BIN to control kernel node creation
which adds support for FIT images without a kernel node when
FIT_LINUX_BIN is set to an empty string.

Extend the functions _get_req_its_paths to return also a list of
not expected paths in the ITS file to verify that no kernel node is
created when FIT_LINUX_BIN = "".

This also adds some basic test coverage for the configuration nodes in
the ITS file.

Minor cleanups:
- Remove unused import of tempfile module.
- Print the configuration for the fitimage.FitImagePyTests when
  overrides are provided.
- Remove empty line at the end of the file.
- Sort the list of bb_vars keys in _test_fitimage_py().

Signed-off-by: Adrian Freihofer <adrian.freihofer@siemens.com>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>